Update imports for moved similarity subpkg
