+++ src/textacy/similarity/__init__.py
@@ -1,1 +1,5 @@
 from . import edits, hybrid, sequences, tokens
+from .edits import hamming, levenshtein, jaro, character_ngrams
+from .hybrid import token_sort_ratio, monge_elkan
+from .sequences import matching_subsequences_ratio
+from .tokens import jaccard, sorensen_dice, tversky, cosine, bag
