+++ tests/similarity/test_edits.py
@@ -1,6 +1,6 @@
 import pytest
 
-from textacy.similarity_ import edits
+from textacy.similarity import edits
 
 
 @pytest.mark.parametrize(
