+++ tests/similarity/test_hybrid.py
@@ -1,6 +1,6 @@
 import pytest
 
-from textacy.similarity_ import hybrid
+from textacy.similarity import hybrid
 
 
 @pytest.mark.parametrize(
